¡THE HORROR‼ *Massive* refactor. Use register file for everything.
This was done as quickly as I could possibly do it. It is very
rough around the edges. $SHOW-ESCAPES currently has no effect.
There is no switch parsing yet, so input and output cannot be
specified. These things should be simple to add now that the
major rework is complete.